moved restore_session to state.lua
