+++ lua/kubectl/actions/commands.lua
@@ -171,21 +171,4 @@ function M.save_config(file_name, data)
   return ok
 end
 
-function M.restore_session()
-  local config = M.load_config("kubectl.json")
-
-  -- change namespace
-  state.ns = config and config.session and config.session.namespace or "All"
-
-  -- change view
-  local session_view = config and config.session and config.session.view or "pods"
-  local ok, view = pcall(require, "kubectl.views." .. string.lower(session_view))
-  if ok then
-    view.View()
-  else
-    local pod_view = require("kubectl.views.pods")
-    pod_view.View()
-  end
-end
-
 return M
